squads/legal-analyst/webapp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pymupdf==1.24.0
pillow==10.4.0
python-jose[cryptography]==3.3.0
websockets==12.0
aiofiles==24.1.0
anthropic==0.34.0
pyyaml==6.0.2
markdown==3.7
python-dotenv==1.0.1
stripe==8.0.0
